requirements.txt
appdirs==1.4.3
click==6.7
cssselect==1.0.1
Jinja2==2.9.5
lxml==3.7.3
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
